requirements.txt
jyotishganit
jupyterlab
pandas
matplotlib
skyfield
ipykernel
jyotichart
fpdf2
svglib
reportlab
